RegistrationTest: move relevant tests from createImageWizard to Registration test
